+++ docs/changelog.md
@@ -8,6 +8,7 @@ This changelog is used to track all major changes to WTFIX.
 
 - Automatically start a new session if no relevant .sid file can be found.
 - Heartbeat monitor now proactively sends heartbeats as well, in compliance with the FIX protocol specification.
+- Pipeline now shuts down gracefully on SIGTERM and SIGINT signals in addition to CTRL+C.
 
 **Fixes**